Familias Profesionales, Ciclos Formativos, Modulos Formativos, Resultados Aprendizaje y Criterios Evaluacion

+++ app/Http/Controllers/API/CicloController.php
@@ -16,13 +16,13 @@ class CicloController extends Controller
     public function index(Request $request)
     {
         $query = CicloFormativo::query();
-        if ($query) {
-            $query->orWhere('nombre', 'like', '%' . $request->q . '%');
+        if ($request->has('search')) {
+            $query->orWhere('nombre', 'like', '%' . $request->search . '%');
         }
 
         return CicloResource::collection(
-            
-            CicloFormativo::orderBy($request->_sort ?? 'id', $request->_order ?? 'asc')
+
+            $query->orderBy($request->_sort ?? 'id', $request->_order ?? 'asc')
                 ->paginate($request->perPage)
         );
     }
@@ -30,8 +30,20 @@ class CicloController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request, $parent_id)
+    public function store(Request $request, $parent_id, CicloFormativo $id)
     {
+
+        $request->validate([
+            'nombre' => 'required|string',
+            'codigo' => 'required|string|unique:ciclos_formativos,codigo',
+            'grado' => 'required|in:G.M.,G.S.,C.E. (G.M.),C.E. (G.S.),BÁSICA,basico,medio,superior',
+            'descripcion' => 'nullable|string',
+        ]);
+
+        abort_if($request->user()->cannot('store', $id), 403, 'No tienes permiso para actualizar este ciclo formativo.');
+
+
+
         $cicloData = json_decode($request->getContent(), true);
 
         $cicloData['familia_profesional_id'] = $parent_id;
@@ -44,7 +56,7 @@ class CicloController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show($parent_id, FamiliaProfesional $familiaProfesional, CicloFormativo $id)
+    public function show($parent_id, CicloFormativo $id)
     {
         return new CicloResource($id);
     }
@@ -54,6 +66,18 @@ class CicloController extends Controller
      */
     public function update(Request $request, $parent_id, CicloFormativo $id)
     {
+
+        abort_if ($request->user()->cannot('update', $id), 403, 'No tienes permiso para actualizar este ciclo formativo.');
+
+
+        $request->validate([
+            'nombre' => 'required|string',
+            'codigo' => 'required|string|unique:ciclos_formativos,codigo,' . $id->id,
+            'grado' => 'required|in:G.M.,G.S.,C.E. (G.M.),C.E. (G.S.),BÁSICA,basico,medio,superior',
+            'descripcion' => 'nullable|string',
+        ]);
+
+
         $cicloData = json_decode($request->getContent(), true);
         $id->update($cicloData);
         return new CicloResource($id);
@@ -62,11 +86,15 @@ class CicloController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($parent_id, CicloFormativo $id)
+    public function destroy(Request $request, $parent_id, CicloFormativo $id)
     {
+        abort_if($request->user()->cannot('delete', $id), 403, 'No tienes permiso para actualizar este ciclo formativo.');
+
+
         try {
+
             $id->delete();
-            return response()->json(null, 204);
+            return response()->json(['message' => 'CicloFormativo eliminado correctamente'], 200);
         } catch (\Exception $e) {
             return response()->json([
                 'message' => 'Error: ' . $e->getMessage()
